% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{DSelector}
\alias{DSelector}
\title{Dantzig Selector}
\usage{
DSelector(X, y, sigma, lambda = 3.5)
}
\arguments{
\item{X}{n * p predictor matrix}

\item{sigma}{sd of noise}

\item{lambda}{regularizing parameter}

\item{Y}{n * 1 vector of observations}
}
\value{
beta \code{beta}
}
\description{
Dantzig selector for sparse estimation
}
\examples{
\dontrun{
a=GSD(n=72,p=256,spa=8)
X=a$X
Y=a$Y
theta=sqrt(8/72)/3
DSelector(X,Y,theta)
}
}
