% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{GD}
\alias{GD}
\title{Gradient descent algorithm}
\usage{
GD(x, y, rate, threshold, max_iter)
}
\arguments{
\item{rate}{the slope of the MSE function}

\item{threshold}{convergence threshold}

\item{max_iter}{maximum number of iterations}

\item{X}{observed data}

\item{Y}{corresponding data}
}
\value{
Optimal intercept \code{c}

Optimal slope \code{m}
}
\description{
use Gradient Descent Algorithm which to find the optimal intercept and gradient for any set of data in which a linear relationship exists.
}
\examples{
\dontrun{
attach(mtcars)
GD(disp, mpg, 0.0000293, 0.001, 2500000)
}
}
