% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{GSD}
\alias{GSD}
\title{Generate Sparse Data}
\usage{
GSD(n, p, spa)
}
\arguments{
\item{n}{the number of row of X (sample size)}

\item{p}{the number of col of X (factor size)}

\item{spa}{the number of nonzero in beta}
}
\value{
X \code{X}

Y \code{Y}
}
\description{
Generate Sparse Data for the use of Dantzig Selector Y=Xbeta+z
}
\examples{
\dontrun{
GSD(n=72,p=256,spa=8)
}
}
