% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Lp}
\alias{Lp}
\title{Standard simplex algorithm}
\usage{
Lp(A, b, cost)
}
\arguments{
\item{A}{coefficient matrix after adding surplus variable}

\item{b}{constant vector}

\item{cost}{cost function parameter vector after adding surplus variable}
}
\value{
Best solver \code{x}
}
\description{
In mathematical optimization, Dantzig's simplex algorithm (or simplex method) is a popular algorithm for linear programming.
}
\examples{
\dontrun{
A = rbind(
c(1,2,-1,0),
c(2,1,0,-1))
b = c(6,6)
cost = c(1,1,0,0)
x <- Lp(A, b, cost)
print(x)
print(cost \%*\% x)
}
}
