% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{Rcpp function}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{standard deviation}

\item{x0}{original value}

\item{N}{length}
}
\value{
generated random numbers \code{x}
}
\description{
an Rcpp function for Exercise 9.4 (page 277, Statistical Computing with R)
}
\examples{
\dontrun{
sd=2
x0=25
N = 2000
rwMetropolis(sigma,x0,N)
}
}
